Refactor UserRepository into focused Health and Achievement interfaces
Extracts health and achievement-related operations from UserRepository into
dedicated HealthRepository and AchievementsRepository interfaces. Updates all
direct call sites in the UI (e.g., AddHealthActivity, MyHealthFragment,
AchievementFragment) to inject the new repository interfaces.

Implementation file splitting is deferred as requested; UserRepositoryImpl
now implements AchievementsRepository, and HealthRepositoryImpl delegates
to UserRepositoryImpl to fulfill the HealthRepository contract.
